added filter, refactored system
